pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "datev-lint"
version = "0.1.0"
description = "DATEV export file validator and linter"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "datev-lint contributors" }
]
keywords = ["datev", "accounting", "validation", "linter", "csv"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Accounting",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0,<3.0",
    "charset-normalizer>=3.0,<4.0",
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.8.0",
    "mypy>=1.13",
    "types-PyYAML>=6.0",
]
rules = [
    "pybloom-live>=4.0",
]
cli = [
    "typer>=0.9.0,<1.0",
    "rich>=13.0,<14.0",
]
pro = [
    "jinja2>=3.0",
    "weasyprint>=60.0",
    "cryptography>=41.0",
]

[project.scripts]
datev-lint = "datev_lint.cli.main:app"

[project.urls]
Homepage = "https://github.com/datev-lint/datev-lint"
Documentation = "https://datev-lint.dev/docs"
Repository = "https://github.com/datev-lint/datev-lint"

# =============================================================================
# Ruff Configuration
# =============================================================================
[tool.ruff]
target-version = "py311"
line-length = 100
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict parameter
]

[tool.ruff.lint.isort]
known-first-party = ["datev_lint"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG001", "S101"]

# =============================================================================
# Mypy Configuration
# =============================================================================
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true

[[tool.mypy.overrides]]
module = [
    "pybloom_live.*",
    "weasyprint.*",
]
ignore_missing_imports = true

# =============================================================================
# Pytest Configuration
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "benchmark: marks tests as performance benchmarks",
]

[tool.coverage.run]
source = ["datev_lint"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
